airflow/requirements.txt
apache-airflow-providers-airbyte==3.2.1
python-dotenv==0.21.1